requirements.txt
aiohttp>=2.0.6
git+https://github.com/Rapptz/discord.py@rewrite
Pillow>=4.0.x
psutil
